requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
cryptography>=3.4.8
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.2.0
flake8==6.0.0
black==23.9.1